Fixed malloc() and free() causing system crashes when accessing PageBitmap
using malloc() or free() and then accessing PageBitmap from the GlobalAllocator would sometimes cause system crashes. moving the function declaration inside of PageFrameAllocator.cpp fixes this issue.
