+++ kernel/src/kernelUtil.cpp
@@ -86,7 +86,7 @@ KernelInfo InitializeKernel(BootInfo* bootInfo){
 
     memset(bootInfo->framebuffer->BaseAddress, 0, bootInfo->framebuffer->BufferSize);
 
-    InitializeHeap((void*)0x0000100000000000, 0x10);
+    InitializeHeap((void*)0x0000100000000000, 0x20);
 
     PrepareInterrupts();
