+++ kernel/src/kernelUtil.h
@@ -4,8 +4,6 @@
 #include "BasicRenderer.h"
 #include "cstr.h"
 #include "efiMemory.h"
-#include "memory.h"
-#include "Bitmap.h"
 #include "paging/PageFrameAllocator.h"
 #include "paging/PageMapIndexer.h"
 #include "paging/paging.h"
