+++ kernel/src/paging/PageFrameAllocator.h
@@ -1,8 +1,18 @@
 #pragma once
 #include "../efiMemory.h"
 #include <stdint.h>
-#include "../Bitmap.h"
 #include "../memory.h"
+#include <stddef.h>
+#include <stdint.h>
+
+class Bitmap{
+    public:
+    size_t Size;
+    uint8_t* Buffer;
+    bool operator[](uint64_t index);
+    bool Set(uint64_t index, bool value);
+    bool Get(uint64_t index);
+};
 
 class PageFrameAllocator {
     public:
